detector/requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
PyYAML==6.0.3
requests==2.33.1
urllib3==2.6.3
flask
psutil==7.2.2
matplotlib=3.10.9